working though MPI basics

+++ assignment-one/mpi-program/common.h
@@ -1,9 +1,10 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <math.h>
 #include <time.h>
 #include <stdbool.h>
 #include <mpi.h>
 
 // Shared Functions
-int *create_one_d_matrix(int size, int *m_out, bool printOutput);
-void printArray(int arr[], int size);
+float *create_one_d_matrix(int size, float *m_out, bool printOutput);
+void printArray(float arr[], int size);
